use parser config for operation naming/filtering and move services processing to preprocessor from parser
