+++ packages/openapi-ts/src/generator/openApi/index.ts
@@ -1,5 +1,6 @@
-import type { Client } from '../../types/client';
-import { OpenApi } from './common/interfaces/OpenApi';
+import type { Client } from './common/interfaces/client';
+import type { Config } from './common/interfaces/config';
+import type { OpenApi } from './common/interfaces/OpenApi';
 import { parse as parseV2 } from './v2/index';
 import { parse as parseV3 } from './v3/index';
 
@@ -11,13 +12,19 @@ export { OpenApi } from './common/interfaces/OpenApi';
  * all the models, services and schema's we should output.
  * @param openApi The OpenAPI spec that we have loaded from disk.
  */
-export function parse(openApi: OpenApi): Client {
+export function parse({
+  openApi,
+  config,
+}: {
+  config: Config;
+  openApi: OpenApi;
+}): Client {
   if ('openapi' in openApi) {
-    return parseV3(openApi);
+    return parseV3({ config, openApi });
   }
 
   if ('swagger' in openApi) {
-    return parseV2(openApi);
+    return parseV2({ config, openApi });
   }
 
   throw new Error(
