+++ packages/openapi-ts/src/generator/openApi/v2/index.ts
@@ -1,26 +1,37 @@
-import type { Client } from '../../../types/client';
+import type { Client } from '../common/interfaces/client';
+import type { Config } from '../common/interfaces/config';
 import { getServiceVersion } from '../common/parser/service';
 import type { OpenApi } from './interfaces/OpenApi';
 import { getModels } from './parser/getModels';
+import { getOperations } from './parser/getOperations';
 import { getServer } from './parser/getServer';
-import { getServices } from './parser/getServices';
 
 /**
  * Parse the OpenAPI specification to a Client model that contains
  * all the models, services and schema's we should output.
  * @param openApi The OpenAPI spec that we have loaded from disk.
  */
-export const parse = (openApi: OpenApi): Client => {
+export const parse = ({
+  openApi,
+  config,
+}: {
+  config: Config;
+  openApi: OpenApi;
+}): Client => {
   const version = getServiceVersion(openApi.info.version);
   const server = getServer(openApi);
   const { models, types } = getModels(openApi);
-  const { operationIds, services } = getServices({ openApi, types });
+  const { operations, operationIds } = getOperations({
+    config,
+    openApi,
+    types,
+  });
 
   return {
     models,
     operationIds,
+    operations,
     server,
-    services,
     types,
     version,
   };
