+++ packages/openapi-ts/src/generator/openApi/v2/parser/operation.ts
@@ -3,11 +3,8 @@ import type {
   Operation,
   OperationParameters,
 } from '../../common/interfaces/client';
-import {
-  getOperationName,
-  getOperationResponseHeader,
-} from '../../common/parser/operation';
-import { getServiceName } from '../../common/parser/service';
+import { Config } from '../../common/interfaces/config';
+import { getOperationResponseHeader } from '../../common/parser/operation';
 import { toSortedByRequired } from '../../common/parser/sort';
 import type { OpenApi } from '../interfaces/OpenApi';
 import type { OpenApiOperation } from '../interfaces/OpenApiOperation';
@@ -20,28 +17,24 @@ export const getOperation = ({
   openApi,
   types,
   pathParams,
-  tag,
   url,
+  config,
 }: {
+  config: Config;
   method: Lowercase<Operation['method']>;
   op: OpenApiOperation;
   openApi: OpenApi;
   pathParams: OperationParameters;
-  tag: string;
   types: Client['types'];
   url: string;
 }): Operation => {
-  const service = getServiceName(tag);
-  const name = getOperationName(url, method, op.operationId);
-
-  const operation: Operation = {
+  const operationWithoutName: Omit<Operation, 'name'> = {
     $refs: [],
     deprecated: op.deprecated === true,
     description: op.description || null,
     id: op.operationId || null,
     imports: [],
     method: method.toUpperCase() as Operation['method'],
-    name,
     parameters: [...pathParams.parameters],
     parametersBody: pathParams.parametersBody,
     parametersCookie: [...pathParams.parametersCookie],
@@ -52,12 +45,17 @@ export const getOperation = ({
     path: url,
     responseHeader: null,
     responses: [],
-    service,
     summary: op.summary || null,
+    tags: op.tags || [],
+  };
+  const operation: Operation = {
+    ...operationWithoutName,
+    name: config.nameFn.operation(operationWithoutName),
   };
 
   if (op.parameters) {
     const parameters = getOperationParameters({
+      config,
       openApi,
       parameters: op.parameters,
       types,
