+++ packages/openapi-ts/src/index.ts
@@ -11,6 +11,12 @@ import type { ClientConfig, Config, UserConfig } from './types/config';
 import { getConfig, isStandaloneClient, setConfig } from './utils/config';
 import { getOpenApiSpec } from './utils/getOpenApiSpec';
 import { registerHandlebarTemplates } from './utils/handlebars';
+import {
+  operationFilterFn,
+  operationNameFn,
+  operationParameterFilterFn,
+  operationParameterNameFn,
+} from './utils/parse';
 import { Performance } from './utils/performance';
 import { postProcessClient } from './utils/postprocess';
 
@@ -319,7 +325,20 @@ export async function createClient(userConfig: UserConfig): Promise<Client[]> {
           >);
 
     Performance.start('parser');
-    const parsed = parse(openApi);
+    const parsed = parse({
+      config: {
+        debug: config.debug,
+        filterFn: {
+          operation: operationFilterFn,
+          operationParameter: operationParameterFilterFn,
+        },
+        nameFn: {
+          operation: operationNameFn,
+          operationParameter: operationParameterNameFn,
+        },
+      },
+      openApi,
+    });
     const client = postProcessClient(parsed);
     Performance.end('parser');
